renamed PowerLawExpCutOff to ExpCutoffPowerLaw (following gammapy), moved test_geometry, in test_utils
