+++ agnpy/tests/test_spectra.py
@@ -2,7 +2,7 @@
 import numpy as np
 import astropy.units as u
 from astropy.constants import m_e
-from agnpy.spectra import PowerLaw, BrokenPowerLaw, LogParabola, PowerLawExpCutOff
+from agnpy.spectra import PowerLaw, BrokenPowerLaw, LogParabola, ExpCutoffPowerLaw
 from agnpy.utils.math import trapz_loglog
 from agnpy.utils.conversion import mec2
 import pytest
@@ -29,7 +29,10 @@ lp_test = LogParabola(
 )
 # global PowerLaw exp Cutoff
 gamma_c_test = 1e3
-pwlec_test = PowerLawExpCutOff(k_e_test, p_test, gamma_c_test, gamma_min_test, gamma_max_test)
+epwl_test = ExpCutoffPowerLaw(
+    k_e_test, p_test, gamma_c_test, gamma_min_test, gamma_max_test
+)
+
 
 def power_law_integral(k_e, p, gamma_min, gamma_max):
     """analytical integral of the power law"""
@@ -346,24 +349,22 @@ class TestLogParabola:
         assert u.isclose(norm, lp(1), atol=0 * u.Unit("cm-3"), rtol=1e-2)
 
 
-
-class TestPowerLawExpCutOff:
-    """class grouping all tests related to the PowerLawExpCutOff spectrum"""
+class TestExpCutoffPowerLaw:
+    """class grouping all tests related to the ExpCutoffPowerLaw spectrum"""
 
     def test_call(self):
         """assert that outside the bounding box the function returns 0"""
         gamma = np.logspace(0, 8)
-        values = pwlec_test(gamma).value
+        values = epwl_test(gamma).value
         condition = (gamma_min_test <= gamma) * (gamma <= gamma_max_test)
         # check that outside the boundaries values are all 0
         assert not np.all(values[~condition])
-
 
     def test_from_normalised_density(self):
         """test the intialisation of the power law from the total particle 
         density"""
         n_e_tot = 1e-5 * u.Unit("cm-3")
-        pwlec = PowerLawExpCutOff.from_normalised_density(
+        epwl = ExpCutoffPowerLaw.from_normalised_density(
             n_e_tot=n_e_tot,
             p=p_test,
             gamma_c=gamma_c_test,
@@ -371,7 +372,7 @@ class TestPowerLawExpCutOff:
             gamma_max=gamma_max_test,
         )
         # calculate n_e_tot
-        n_e_tot_calc = pwlec.integral(
+        n_e_tot_calc = epwl.integral(
             gamma_low=gamma_min_test, gamma_up=gamma_max_test, gamma_power=0
         )
         assert u.isclose(n_e_tot, n_e_tot_calc, atol=0 * u.Unit("cm-3"), rtol=1e-2)
@@ -380,12 +381,17 @@ class TestPowerLawExpCutOff:
         """test the intialisation of the power law from the total particle 
         energy density"""
         u_e = 3e-4 * u.Unit("erg cm-3")
-        pwlec = PowerLawExpCutOff.from_normalised_energy_density(
-            u_e=u_e, p=p_test, gamma_c=gamma_c_test, gamma_min=gamma_min_test, gamma_max=gamma_max_test
+        epwl = ExpCutoffPowerLaw.from_normalised_energy_density(
+            u_e=u_e,
+            p=p_test,
+            gamma_c=gamma_c_test,
+            gamma_min=gamma_min_test,
+            gamma_max=gamma_max_test,
         )
         # calculate u_e
-        u_e_calc = mec2 * pwlec.integral(
-            gamma_low=gamma_min_test, gamma_up=gamma_max_test, gamma_power=1)
+        u_e_calc = mec2 * epwl.integral(
+            gamma_low=gamma_min_test, gamma_up=gamma_max_test, gamma_power=1
+        )
         print(u_e)
         print(u_e_calc)
         assert u.isclose(u_e, u_e_calc, atol=0 * u.Unit("erg cm-3"), rtol=1e-2)
@@ -394,7 +400,11 @@ class TestPowerLawExpCutOff:
         """test the intialisation of the powerlaw from the normalisation at 
         gamma = 1"""
         norm = 1e-13 * u.Unit("cm-3")
-        pwlec = PowerLawExpCutOff.from_norm_at_gamma_1(
-            norm=norm, p=p_test, gamma_c=gamma_c_test, gamma_min=1, gamma_max=gamma_max_test
+        epwl = ExpCutoffPowerLaw.from_norm_at_gamma_1(
+            norm=norm,
+            p=p_test,
+            gamma_c=gamma_c_test,
+            gamma_min=1,
+            gamma_max=gamma_max_test,
         )
-        assert u.isclose(norm, pwlec(1), atol=0 * u.Unit("cm-3"), rtol=1e-2)
+        assert u.isclose(norm, epwl(1), atol=0 * u.Unit("cm-3"), rtol=1e-2)
